Fixed WeaponStances. Created Spawner, TargetEnemy.
WeaponStances had not been working as expected because the XY had been baked into the pose erroneously.
Created a Spawner script to spawn a gameobject either once, continuously, or to maintain a certain number of spawns. Created the TargetEnemy as a temp for enemies. Works.
Still having issue where as soon as play begins, the avatar is no longer facing the same direction as the root transform.