Root transform now correctly updates to match character hips
All idle animations rotations are now based on character orientation for purposes of root motion. Now, the hips turn over time to face where the mouse is pointing, and the character's root transform matches the hips.
Adjusted the weapon container to be on the right shoulder, and moved/rotated it to better match the aim of the mouse. Also tweaked the transform of the visuals of the handgun for the same reason.