Players and Enemies now ragdoll when killed.
Everything works fine, but the canvas says to remove the NavMeshAgent on enemies when killed, and the Player script from the Player when killed. This causes problems, because many other scripts that are NOT destroyed rely on those. Even with the errors from trying to do so, everything looks fine in-game, though. Hopefully this will be addressed in the video I have yet to watch.