Enemies now drop loot when killed.
Spawner script is now on enemy as well, and adjusted to have an array of prefabs to choose from when spawning, as when as logic for whether to always pick the first item or choose randomly. Enemies now choose a random pickup from the list, which is includes all weapons and first-aid pickups, and drops one on death.
I also fixed the issue with destroying certain monobehaviors on death once I realized that a function can destroy its own monobehavior, THEN others that it relies on, all in a row within the same update.