Reconfigured HealthBar for player. Enemies have health bars above heads.
Added a UIManager, which now manages the player's and the enemies's health bars, adding listeners to the characters' health scripts to update the health bars when appropriate. Each enemy has a World-Space Canvas on them that holds their health bar over their head. UI Camera added to show only UI, and on top of any of the non-UI stuff shown by the main camera.