UI screens added. Main menu scene. Can reload level, quit. Score added.
The Singleton parent class removed from GM and UIM, replaced with simple Awake() logic to prevent errors when reloading the level. Win, Lose, and Pause screens added, all handled by the UIManager. Score added, consisting of a base score for the enemy type and a modifier via the weapon type.