Settings Menu implemented, but not complete
Settings Menu is now a prefab with its own script that works, can be opened as a submenu of any other menu without breaking that level. Everything is in place, but there is currently no audio to actually change the volume of, and the resolution, graphics settings, and fullscreen/windowed mode are not yet implemented either. Once they are, it will be simple to finish the Settings.