Ammo count on UI now updates when new weapon equipped
The call to update the UI was happening after the call to instantiate the new weapon into the Player's hands, but before that weapon had finished setting itself up, so it was updating the screen to 0 rounds left. Now, the update is called at the start of the weapon as well, after it is set up.